Cast reservation flags before sending to MySQL.
Flags are stored in a smallint, which can only hold the first 16 bits
worth out of 32 bits of flags currently in use.

MySQL's overflow rules will treat any value > 0xffff as 0xffff, rather than
dropping the higher-order bits (flags), which means the stored value not
only loses the higher-order bits but corrupts the lower-order as well.

The 19.05 release extends the column to bigint (64 bit).

Bug 6969.

+++ src/plugins/accounting_storage/mysql/as_mysql_resv.c
@@ -93,8 +93,8 @@ static int _setup_resv_limits(slurmdb_reservation_rec_t *resv,
 
 	if (resv->flags != NO_VAL) {
 		xstrcat(*cols, ", flags");
-		xstrfmtcat(*vals, ", %u", resv->flags);
-		xstrfmtcat(*extra, ", flags=%u", resv->flags);
+		xstrfmtcat(*vals, ", %u", (uint16_t)resv->flags);
+		xstrfmtcat(*extra, ", flags=%u", (uint16_t)resv->flags);
 	}
 
 	if (resv->name) {
